refactor: 전시/작품 카드 공통 컴포넌트화 및 불필요 렌더링 최적화
