+++ src/components/mypage/ArtworkCard.tsx
@@ -1,10 +1,8 @@
-import { useState } from 'react';
+import { Bookmark } from 'lucide-react';
 
-import { Bookmark, Pencil } from 'lucide-react';
-
-import { useMemoPolicy } from '@/hooks/usePolicy';
 import type { SavedArtworkItem } from '@/types/mypage';
-import { hasPermission } from '@/utils/hasPermission';
+
+import { MemoFooter } from './MemoFooter';
 
 interface ArtworkCardProps {
   item: SavedArtworkItem;
@@ -23,14 +21,6 @@ export function ArtworkCard({
   onDeleteMemo,
   onOpen,
 }: ArtworkCardProps) {
-  const hasMemo = Boolean(item.memo);
-  const memoPolicy = useMemoPolicy(item.userId === undefined ? undefined : { userId: item.userId });
-  const canViewMemo = hasPermission(memoPolicy, 'view');
-  const canUpsertMemo = hasPermission(memoPolicy, 'upsert');
-  const canDeleteMemo = hasPermission(memoPolicy, 'delete');
-  const [isEditingMemo, setIsEditingMemo] = useState(false);
-  const [memoInput, setMemoInput] = useState(item.memo ?? '');
-
   /* 카드 영역은 div라서 키보드로도 열 수 있도록 역할과 키 처리를 함께 부여합니다. */
   const openHandlers = onOpen
     ? {
@@ -45,41 +35,6 @@ export function ArtworkCard({
         },
       }
     : {};
-
-  const handleCancelMemo = () => {
-    setMemoInput(item.memo ?? '');
-    setIsEditingMemo(false);
-  };
-
-  const handleStartMemoEdit = () => {
-    if (!canUpsertMemo) return;
-
-    setMemoInput(item.memo ?? '');
-    setIsEditingMemo(true);
-  };
-
-  const handleSaveMemo = () => {
-    if (!canUpsertMemo) return;
-
-    const trimmedInput = memoInput.trim();
-    const originalMemo = item.memo ?? '';
-
-    if (trimmedInput === '') {
-      if (hasMemo && canDeleteMemo) {
-        onDeleteMemo?.(item);
-      }
-    } else if (trimmedInput !== originalMemo) {
-      onSaveMemo?.(item, trimmedInput);
-    }
-    setIsEditingMemo(false);
-  };
-
-  const handleDeleteMemo = () => {
-    if (!canDeleteMemo) return;
-
-    onDeleteMemo?.(item);
-  };
-
   return (
     <article className="bg-card rounded-2xl shadow-[8px_8px_18px_0px_rgba(67,0,209,0.04)] flex flex-col overflow-hidden">
       <div {...openHandlers} className={onOpen ? 'cursor-pointer px-1.5 py-2.5' : 'px-1.5 py-2.5'}>
@@ -117,66 +72,14 @@ export function ArtworkCard({
         <div className="typo-body-xs-regular text-main truncate">{item.artist}</div>
       </div>
 
-      {!isArtistView && canViewMemo && (
-        <footer className="px-2.5 py-2 bg-box200 flex flex-col justify-center">
-          {isEditingMemo ? (
-            <div className="self-stretch flex flex-col gap-[6px]">
-              <div className="self-stretch flex items-center justify-between">
-                <div className="flex items-center gap-1.5 text-faint typo-body-xs-regular">
-                  <Pencil className="size-3 shrink-0" />
-                  <span>메모</span>
-                </div>
-                <button
-                  type="button"
-                  className="typo-body-xs-regular text-faint underline underline-offset-2"
-                  onMouseDown={(e) => {
-                    e.preventDefault();
-                    handleSaveMemo();
-                  }}
-                >
-                  확인
-                </button>
-              </div>
-              <textarea
-                value={memoInput}
-                onChange={(event) => setMemoInput(event.target.value)}
-                onBlur={handleSaveMemo}
-                placeholder=""
-                className="w-full resize-none bg-transparent typo-body-xs-regular text-main outline-none"
-                autoFocus
-                rows={1}
-              />
-            </div>
-          ) : hasMemo ? (
-            <div className="self-stretch flex items-start gap-2">
-              <button
-                type="button"
-                className="min-w-0 flex-1 text-left typo-body-xs-regular text-faint line-clamp-2"
-                onClick={handleStartMemoEdit}
-              >
-                {item.memo}
-              </button>
-              {canDeleteMemo && (
-                <button
-                  type="button"
-                  className="shrink-0 typo-body-xs-regular text-faint underline"
-                  onClick={handleDeleteMemo}
-                >
-                  삭제
-                </button>
-              )}
-            </div>
-          ) : canUpsertMemo ? (
-            <button
-              type="button"
-              className="self-stretch flex items-center gap-1.5 text-left typo-body-xs-regular text-faint"
-              onClick={handleStartMemoEdit}
-            >
-              <Pencil className="size-3 shrink-0 text-faint" />
-              <span>메모</span>
-            </button>
-          ) : null}
-        </footer>
+      {!isArtistView && (
+        <MemoFooter
+          className="px-2.5 py-2"
+          memo={item.memo}
+          userId={item.userId}
+          onSave={(memo) => onSaveMemo?.(item, memo)}
+          onDelete={() => onDeleteMemo?.(item)}
+        />
       )}
     </article>
   );
