+++ src/components/mypage/index.ts
@@ -2,5 +2,6 @@ export { ArtistCard } from './ArtistCard';
 export { ArtworkCard } from './ArtworkCard';
 export { AuthPageHeader } from './AuthPageHeader';
 export { ExhibitionCard } from './ExhibitionCard';
+export { MemoFooter } from './MemoFooter';
 export { MyPageHeader } from './MyPageHeader';
 export { SettingsSheet } from './SettingsSheet';
